rename services_serializers, update tests, update view, rename services_views
